Ensure that root is black to prevent rebalancing issues (#73)
We've found another issue with rebalancing when there is a one element tree. When creating a tree with two elements, the root is black and child is red. However, upon removal of the root, the child is promoted but not set to black. Later on, when another element is pushed, this allows for a red node to have a red child.

The fix is to simply set the root to always be black.

* root must always be black

* add test case

* trailing EOL

* update package.json

Co-authored-by: Thomas Tay <Tay.Thomas@microsoft.com>